S302T200-143 #time 1h Debugged the Cucumber tests.
Removed the use of MockMvc because MockMvc.perform() kept giving null pointer exceptions. Now calling the controller method directly but this keeps throwing unexpected exception "null". Removed the Cucumber test for AC8 because I can't write a test for this without MockMvc, and added AC8 to the spreadsheet of manual tests.

?test ?fix ?document